requirements.txt
docx2txt==0.8
langchain==0.0.225
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-openai==0.1.8
pandas
openpyxl
xlrd
# for images
# pillow 
# easyocr